RemoveRZandZbeforeMeasure -> RemoveDiagonalGatesBeforeMeasure
