+++ qiskit/transpiler/passes/__init__.py
@@ -37,6 +37,6 @@ from .mapping.dense_layout import DenseLayout
 from .mapping.noise_adaptive_layout import NoiseAdaptiveLayout
 from .mapping.basic_swap import BasicSwap
 from .mapping.lookahead_swap import LookaheadSwap
-from .remove_rz_and_z_before_measure import RemoveRZandZbeforeMeasure
+from .remove_diagonal_gates_before_measurere import RemoveDiagonalGatesBeforeMeasure
 from .mapping.stochastic_swap import StochasticSwap
 from .mapping.legacy_swap import LegacySwap
